requirements.txt
flask
pymongo[srv]
bson
python-dotenv
gunicorn